lab2/backend/requirements.txt
# FastAPI and Server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Database - PostgreSQL with pgvector
psycopg[binary,pool]>=3.2.0
pgvector>=0.2.5

# AWS SDK
boto3>=1.35.0
botocore>=1.35.0

# Configuration and Environment
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Core Dependencies
numpy>=1.22.0,<2.0.0
tenacity>=8.0.0
rich>=13.0.0

# Logging and Utilities
python-json-logger>=3.0.0

# Lab 2: Agents and MCP
strands-agents
strands-agents-tools
strands-agents-builder